Gateとグラフようデータの作成

+++ laravel-admin/app/Http/Controllers/RoleController.php
@@ -12,15 +12,19 @@ class RoleController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view', 'roles');
+
         return RoleResource::collection(Role::all());
     }
 
     public function store(Request $request)
     {
+        Gate::authorize('edit', 'roles');
+
         $role = Role::create($request->only('name'));
 
-        if($permissions = $request->input('permissions')){
-            foreach($permissions as $permission_id) {
+        if ($permissions = $request->input('permissions')) {
+            foreach ($permissions as $permission_id) {
                 DB::table('role_permission')->insert([
                     'role_id'=>$role->id,
                     'permission_id'=>$permission_id,
@@ -33,18 +37,22 @@ class RoleController extends Controller
 
     public function show($id)
     {
+        Gate::authorize('view', 'roles');
+
         return new RoleResource(Role::find($id));
     }
 
     public function update(Request $request, $id)
     {
+        Gate::authorize('edit', 'roles');
+
         $role=Role::find($id);
         $role->update($request->only('name'));
 
-        DB::table('role_permission')->where('role_id',$role->id)->delete();
+        DB::table('role_permission')->where('role_id', $role->id)->delete();
 
-        if($permissions = $request->input('permissions')){
-            foreach($permissions as $permission_id) {
+        if ($permissions = $request->input('permissions')) {
+            foreach ($permissions as $permission_id) {
                 DB::table('role_permission')->insert([
                     'role_id'=>$role->id,
                     'permission_id'=>$permission_id,
@@ -57,7 +65,9 @@ class RoleController extends Controller
 
     public function destroy($id)
     {
-        DB::table('role_permission')->where('role_id',$id)->delete();
+        Gate::authorize('edit', 'roles');
+
+        DB::table('role_permission')->where('role_id', $id)->delete();
 
         Role::destroy($id);
